+++ TimberCommons/package/CHANGELOG.md
@@ -1,4 +1,5 @@
 # v1.7 (started on 3/1/2024):
+* [Fix #37] Water tower with zero coverage consumes water.
 * [Fix #43] Water towers override the highlighted range of other buildings.
 * [Fix #45] Mechanical towers don't properly update the effective range.
 * [Enhancement] Improved powered irrigation towers performance when the Power Grid gets fluctuations.